Fix temp files

+++ internal/filemanager/utils.go
@@ -163,6 +163,6 @@ func GetAvailableSpaceGB(path string) (float64, error) {
 		logutils.Log.WithError(err).Error("Failed to get filesystem stats")
 		return 0, err
 	}
-	availableSpaceGB := float64(int64(stat.Bavail)*stat.Bsize) / (1024 * 1024 * 1024) // #nosec G115
+	availableSpaceGB := float64(int64(stat.Bavail)*int64(stat.Bsize)) / (1024 * 1024 * 1024) // #nosec G115
 	return availableSpaceGB, nil
 }
